fix: some BFF rafactoring
