+++ packages/bff/src/data-source.ts
@@ -2,23 +2,18 @@ import 'reflect-metadata';
 import { DataSource, DataSourceOptions } from 'typeorm';
 import config from './config';
 
-const isDev = process.env.DEV_ENV === 'dev';
-
-console.log('isDev: ', isDev);
-
 console.log(
-  'REMINDER: In datasource file, synchronize needs to be changed to false for production',
-  process.env.DEV_ENV,
+  `REMINDER: In datasource file, synchronize needs to be changed to false for production (isDev: ${config.isDev}`,
 );
 
 export const connectionOptions: DataSourceOptions = {
   type: 'postgres',
   url: config.postgresql.connectionString,
-  synchronize: isDev,
-  logging: isDev,
+  synchronize: config.isDev,
+  logging: false,
   entities: ['src/entities/*{.ts,.js}'],
   migrations: [__dirname + '/migrations/**/*.ts'],
-  ...(!isDev && {
+  ...(!config.isDev && {
     extra: {
       ssl: {
         rejectUnauthorized: false,
