+++ packages/bff/src/oidc/cookies.ts
@@ -1,8 +1,11 @@
 import { CookieOptions } from 'express';
 import session from 'express-session';
 
+// TODO: refactor to ../config.ts
+const cookieName = process.env.COOKIE_NAME || 'cookieName';
+const secret = process.env.SESSION_SECRET || 'SecretHere'
+
 export const setCookie = (res: any, value: string) => {
-  const cookieName = process.env.COOKIE_NAME || 'cookieName';
   const options: CookieOptions = {
     httpOnly: true, // Cookie not accessible via client-side script
     secure: process.env.ENABLE_HTTPS === 'true' ? true : false, // Cookie will be sent only over HTTPS if set to true
@@ -11,13 +14,12 @@ export const setCookie = (res: any, value: string) => {
 };
 
 export const readCookie = (req: any) => {
-  const cookieName = process.env.COOKIE_NAME || 'cookieName';
   return req.cookies[cookieName] || null;
 };
 
 export const deleteCookie = async (res: any) => {
   try {
-    res.clearCookie[process.env.COOKIE_NAME || 'cookieName'];
+    res.clearCookie[cookieName];
     res.clearCookie('connect.sid');
   } catch (error) {
     console.error('deleteCookie failed: ', error);
@@ -25,7 +27,7 @@ export const deleteCookie = async (res: any) => {
 };
 
 export const sessionMiddleware = session({
-  secret: process.env.SESSION_SECRET || 'SecretHere',
+  secret,
   resave: false,
   saveUninitialized: false,
   cookie: { maxAge: 30 * 24 * 60 * 60 * 1000, secure: process.env.ENABLE_HTTPS ? true : false },
