--- packages/bff/src/routes/index.ts
@@ -1,8 +1,0 @@
-import express from 'express';
-import { authenticationRouter } from './authenticationRouter';
-
-const routes = express.Router();
-
-routes.use('/auth', authenticationRouter);
-
-export { routes };
